[Reorg] Undo commands to own file
