+++ src/model/nifdelegate.cpp
@@ -37,6 +37,7 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "nifmodel.h"
 #include "nifproxy.h"
 #include "spellbook.h"
+#include "undocommands.h"
 #include "widgets/valueedit.h"
 #include "widgets/nifcheckboxlist.h"
