+++ src/model/nifmodel.cpp
@@ -3049,89 +3049,3 @@ QVariant NifModelEval::operator()( const QVariant & v ) const
 
 	return v;
 }
-
-
-/*
- *  ChangeValueCommand
- */
-
-ChangeValueCommand::ChangeValueCommand( const QModelIndex & index,
-	const QVariant & value, const QString & valueString, const QString & valueType, NifModel * model )
-	: QUndoCommand(), nif( model ), idx( index )
-{
-	oldValue = index.data( Qt::EditRole );
-	newValue = value;
-
-	auto oldTxt = index.data( Qt::DisplayRole ).toString();
-	auto newTxt = valueString;
-
-	if ( !newTxt.isEmpty() )
-		setText( QApplication::translate( "ChangeValueCommand", "Set %1 to %2" ).arg( valueType ).arg( newTxt ) );
-	else
-		setText( QApplication::translate( "ChangeValueCommand", "Modify %1" ).arg( valueType ) );
-}
-
-ChangeValueCommand::ChangeValueCommand( const QModelIndex & index, const NifValue & oldVal, 
-										const NifValue & newVal, const QString & valueType, NifModel * model )
-	: QUndoCommand(), nif( model ), idx( index )
-{
-	oldValue = oldVal.toVariant();
-	newValue = newVal.toVariant();
-
-	auto oldTxt = oldVal.toString();
-	auto newTxt = newVal.toString();
-
-	if ( !newTxt.isEmpty() )
-		setText( QApplication::translate( "ChangeValueCommand", "Set %1 to %2" ).arg( valueType ).arg( newTxt ) );
-	else
-		setText( QApplication::translate( "ChangeValueCommand", "Modify %1" ).arg( valueType ) );
-}
-
-void ChangeValueCommand::redo()
-{
-	//qDebug() << "Redoing";
-	nif->setData( idx, newValue, Qt::EditRole );
-
-	//qDebug() << nif->data( idx ).toString();
-}
-
-void ChangeValueCommand::undo()
-{
-	//qDebug() << "Undoing";
-	nif->setData( idx, oldValue, Qt::EditRole );
-
-	//qDebug() << nif->data( idx ).toString();
-}
-
-
-/*
- *  ToggleCheckBoxListCommand
- */
-
-ToggleCheckBoxListCommand::ToggleCheckBoxListCommand( const QModelIndex & index,
-	const QVariant & value, const QString & valueType, NifModel * model )
-	: QUndoCommand(), nif( model ), idx( index )
-{
-	oldValue = index.data( Qt::EditRole );
-	newValue = value;
-
-	auto oldTxt = index.data( Qt::DisplayRole ).toString();
-
-	setText( QApplication::translate( "ToggleCheckBoxListCommand", "Modify %1" ).arg( valueType ) );
-}
-
-void ToggleCheckBoxListCommand::redo()
-{
-	//qDebug() << "Redoing";
-	nif->setData( idx, newValue, Qt::EditRole );
-
-	//qDebug() << nif->data( idx ).toString();
-}
-
-void ToggleCheckBoxListCommand::undo()
-{
-	//qDebug() << "Undoing";
-	nif->setData( idx, oldValue, Qt::EditRole );
-
-	//qDebug() << nif->data( idx ).toString();
-}
